# packages/sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "items": {
                  "additionalProperties": false,
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ],
                  "type": "object"
                },
                "type": "array"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "LocalMail API key using the lm_live_… or lm_admin_… prefix convention.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local-first email inbox API for AI agents.",
    "title": "LocalMail API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/search": {
      "get": {
        "operationId": "searchMessages",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "fts",
              "enum": [
                "fts",
                "semantic"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inbox_id",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "created_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "direction": {
                            "enum": [
                              "inbound",
                              "outbound"
                            ],
                            "type": "string"
                          },
                          "from": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "jev_decision": {
                            "additionalProperties": false,
                            "nullable": true,
                            "properties": {
                              "answers": {
                                "additionalProperties": {},
                                "type": "object"
                              },
                              "created_at": {
                                "format": "date-time",
                                "type": "string"
                              },
                              "latency_ms": {
                                "type": "integer"
                              }
                            },
                            "required": [
                              "answers",
                              "latency_ms",
                              "created_at"
                            ],
                            "type": "object"
                          },
                          "labels": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "mode": {
                            "enum": [
                              "fts",
                              "semantic"
                            ],
                            "type": "string"
                          },
                          "preview": {
                            "nullable": true,
                            "type": "string"
                          },
                          "rank": {
                            "type": "number"
                          },
                          "received_at": {
                            "format": "date-time",
                            "nullable": true,
                            "type": "string"
                          },
                          "subject": {
                            "nullable": true,
                            "type": "string"
                          },
                          "thread_id": {
                            "type": "string"
                          },
                          "to": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          }
                        },
                        "required": [
                          "id",
                          "thread_id",
                          "subject",
                          "preview",
                          "labels",
                          "direction",
                          "from",
                          "to",
                          "received_at",
                          "created_at",
                          "rank",
                          "mode"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "next_page_token": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "data",
                    "next_page_token"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Search"
        ]
      }
    },
...
  },
...
}